Factor out useHoveringPath and add to other drop components.
