+++ src/components/Subthoughts/SubthoughtsDropEmpty.tsx
@@ -1,21 +1,19 @@
 import classNames from 'classnames'
-import React, { useEffect } from 'react'
+import React from 'react'
 import { ConnectDropTarget } from 'react-dnd'
 import { useSelector } from 'react-redux'
-import DragThoughtZone from '../../@types/DragThoughtZone'
 import DropThoughtZone from '../../@types/DropThoughtZone'
 import SimplePath from '../../@types/SimplePath'
 import State from '../../@types/State'
-import dragInProgressActionCreator from '../../action-creators/dragInProgress'
 import globals from '../../globals'
 import getThoughtById from '../../selectors/getThoughtById'
-import store from '../../stores/app'
 import equalPath from '../../util/equalPath'
 import head from '../../util/head'
 import isDivider from '../../util/isDivider'
 import strip from '../../util/strip'
 import DragAndDropSubthoughts from '../DragAndDropSubthoughts'
 import useDropHoverColor from './useDropHoverColor'
+import useHoveringPath from './useHoveringPath'
 
 /** A drop target when there are no children or the thought is collapsed. The drop-hover components are ThoughtDropHover, SubthoughtsDropEnd, and SubthoughtsDropEmpty. */
 const SubthoughtsDropEmpty = ({
@@ -35,22 +33,7 @@ const SubthoughtsDropEmpty = ({
   const draggingThought = useSelector((state: State) => state.draggingThought)
   const value = useSelector((state: State) => getThoughtById(state, head(simplePath))?.value || '')
   const dropHoverColor = useDropHoverColor(depth || 0)
-
-  // When SubthoughtsDropEmpty is hovered over during drag, update the hoveringPath.
-  // Check dragInProgress to ensure the drag has not been aborted (e.g. by shaking).
-  useEffect(() => {
-    if (isHovering && store.getState().dragInProgress) {
-      store.dispatch(
-        dragInProgressActionCreator({
-          value: true,
-          draggingThought: store.getState().draggingThought,
-          hoveringPath: simplePath,
-          hoverZone: DropThoughtZone.SubthoughtsDrop,
-          sourceZone: DragThoughtZone.Thoughts,
-        }),
-      )
-    }
-  }, [isHovering])
+  useHoveringPath(simplePath, !!isHovering, DropThoughtZone.SubthoughtsDrop)
 
   // Why do we bail if the thought is being dragged?
   // Even though canDrop will prevent a thought from being dropped on itself, we still should prevent rendering the drop target at all, otherwise it will obscure valid drop targets.
